Rename BasicAuthFilters to SunriseHttpFilters, as it is a global thing, and move inside `play` package
