+++ common/app/com/commercetools/sunrise/play/SunriseHttpFilters.java
@@ -1,5 +1,6 @@
-package com.commercetools.sunrise.common.basicauth;
+package com.commercetools.sunrise.play;
 
+import com.commercetools.sunrise.common.basicauth.BasicAuthFilter;
 import play.filters.csrf.CSRFFilter;
 import play.http.HttpFilters;
 import play.mvc.EssentialFilter;
@@ -10,12 +11,12 @@ import javax.inject.Inject;
  * Based on play.mvc.http.DefaultHttpFilters with fix https://github.com/playframework/playframework/pull/6238, which will be released with Play 2.5.5.
  * REMOVE AS SOON AS IT IS RELEASED!
  */
-public class BasicAuthHttpFilters implements HttpFilters{
+public class SunriseHttpFilters implements HttpFilters {
 
     private final EssentialFilter[] filters;
 
     @Inject
-    public BasicAuthHttpFilters(final BasicAuthFilter basicAuthFilter, final CSRFFilter csrfFilter) {
+    public SunriseHttpFilters(final BasicAuthFilter basicAuthFilter, final CSRFFilter csrfFilter) {
         this.filters = new EssentialFilter[] {basicAuthFilter.asJava(), csrfFilter.asJava()};
     }
