+++ common/test/com/commercetools/sunrise/common/utils/SunriseConfigurationTest.java
@@ -1,6 +1,6 @@
 package com.commercetools.sunrise.common.utils;
 
-import com.commercetools.sunrise.common.configuration.SunriseConfiguration;
+import com.commercetools.sunrise.play.SunriseConfiguration;
 import org.junit.Test;
 import play.Configuration;
