+++ conf/application.conf
@@ -314,8 +314,8 @@ play.crypto.secret=${?APPLICATION_SECRET}
 
 # INJECTION
 # ~~~~~
-play.application.loader = "com.commercetools.sunrise.common.configuration.SunriseApplicationLoader"
-play.http.filters = "com.commercetools.sunrise.common.basicauth.BasicAuthHttpFilters"
+play.application.loader = "com.commercetools.sunrise.play.SunriseApplicationLoader"
+play.http.filters = "com.commercetools.sunrise.play.SunriseHttpFilters"
 
 play.modules.enabled += "com.commercetools.sunrise.common.contexts.RequestScopeModule"
 play.modules.enabled += "demo.common.ComponentsModule"
